feat(n): extract DashboardTab, LeadsTab, AnalyticsTab with next/dynamic
The advisor portal page.tsx was a monolithic 2,620-line "use client"
component that shipped the full JS for all 8 tabs to the browser on
every page load, regardless of which tab the advisor visited.

Three of the heaviest tabs — Dashboard (~365 LOC of JSX), Leads (~275
LOC) and Analytics (~120 LOC) — are now extracted into dedicated "use
client" components that next/dynamic lazy-imports. The browser only
fetches the JS for a tab when the advisor actually navigates to it,
cutting the initial interactive payload.

Implementation notes:
- Shared types (Advisor, Lead, Stats, ViewType, CategoryPricing,
  DisputeModal, etc.) extracted to app/advisor-portal/types.ts so
  page.tsx and all tab files can import from a single source of truth
  without circular deps.
- DashboardTab receives all read-only state (advisor, stats, leads,
  reviews, charts, profileCompleteness) plus two callbacks
  (onNavigate, onDismissOnboarding). No fetch side-effects — parent
  owns all data.
- LeadsTab receives leads array + all filter state + onLeadSortByQualityChange /
  onHotLeadsOnlyChange (bool setters, not toggles, so the "Clear
  filters" button can reset to false without toggling an already-false
  value). Dispute-modal state stays in page.tsx; the tab fires
  onOpenDisputeModal with the leadId/daysLeft tuple.
- AnalyticsTab is the simplest: read-only stats + leads funnel counts
  + onNavigate for the "Write one →" article link.
- page.tsx shrinks from 2,620 → 1,847 lines (−773 LOC).

Verified callers: these components are only imported by
app/advisor-portal/page.tsx via next/dynamic. No other callers.

Idempotency / safety: pure code movement — zero business-logic
changes. Behaviour is identical to before: the same JSX, handlers,
and data shapes, just loaded on demand per tab rather than upfront.
Props are TypeScript-typed; any future drift between parent state and
tab props is caught at CI tsc time.

Rollback: revert this commit. page.tsx regains the inline JSX.

Queue: N-03b (stream N, iter 38)
Refs: #242
Audit: docs/audits/codebase-health-2026-04-24.md §6 (N stream)
